Putting the utility functions into a separate file. Modified : Putting the support functions of the main mod in a separate file.
