+++ control.lua
@@ -1,25 +1,5 @@
 require("util")
--- Utility functions
-
-function roundn(x)
-  return x + 0.5 - (x + 0.5) % 1
-end
-
-function inrange(position)
-  return ((position[1]-myplayer.position.x)^2+(position[2]-myplayer.position.y)^2) < 36
-end
-
-function debugprint(msg)
-	if settings.global["tas-verbose-logging"].value then 
-		outp = myplayer or game
-		outp.print("[" .. game.tick .. "] " .. msg)
-	end
-end
-
-function errprint(msg)
-	outp = myplayer or game
-	outp.print("[" .. game.tick .. "]  ___WARNING___ " .. msg)
-end
+require("utility_functions")
 
 -- Global variables initialization
 myplayer = nil
